start on fixing tests
